Safari-611.1.10/LayoutTests/inspector/canvas/requestContent-webgl2-expected.txt

Test that CanvasAgent.requestContent can retrieve a dataURL with the current content of the WebGL2 canvas.


== Running test suite: Canvas.requestContentWebGL2
-- Running test case: Canvas.requestContentWebGL2.validCanvasId
PASS: Content should not be empty.
